[frontend] Renaming
