+++ opencti-platform/opencti-front/src/private/components/settings/Root.jsx
@@ -28,7 +28,7 @@ import Rules from './Rules';
 import Sessions from './Sessions';
 import Settings from './Settings';
 import SettingsOrganizations from './SettingsOrganizations';
-import FileIndexingConfiguration from './file_indexing/FileIndexingConfiguration';
+import FileIndexing from './file_indexing/FileIndexing';
 import StatusTemplates from './status_templates/StatusTemplates';
 import RootSubType from './sub_types/Root';
 import SubTypes from './sub_types/SubTypes';
@@ -186,7 +186,7 @@ const Root = () => {
         />
         <BoundaryRoute
           path="/dashboard/settings/file_indexing"
-          render={() => <FileIndexingConfiguration />}
+          render={() => <FileIndexing />}
         />
         <BoundaryRoute
           exact
